More infrastructure
Added code to generate and stitch together plots that represent an entire episode of acting greedy with respect to the Q matrix. Also added a plot of stiffness and phidot over the same episode. Possible issue: stiffness and phidot don't seem to change over the final episode. This needs investigation.