Removed phi[t] NDSolve
Removed Lagranges equation from the set of equations for which NDSolve finds solutions for. In place of this, phi[t],phi'[t], and phi''[t] are specified as replacement rules for the ODEs. Created function phiUpdate which computes phi based on action taken and previous phi. This isn't quite working yet. Need to find a way to store phi[t] and phi'[t] in lists that are the same "length" as the interpolating functions output by NDSolve so phi[t] can be plotted and used for animations.